+++ addOns/authhelper/CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Allow to use zero login page wait for Client Script and Browser Based Authentication methods through the GUI.
 - Ensure Client Script Based Authentication method has a clean state when reauthenticating.
 - Handle missing username field in Browser Based Authentication.
+- Correct the processing of cookies with the same name in Header Based Session Management method.
 
 ## [0.25.0] - 2025-03-25
 ### Changed